CON-189: Move validateBlockCheckpoint to test since it was only used in one place in the main code
